\name{section.ids.for.unit}
\alias{section.ids.for.unit}
\title{Lists the sections in a unit}
\usage{
  section.ids.for.unit(unit.id, toc)
}
\arguments{
  \item{unit.id}{the unit ID from which the sections are to
  be draw}

  \item{toc}{the TOC object to query}
}
\value{
  a list of the unique section IDs in the TOC within the
  specified unit
}
\description{
  This function lists the unique section IDs in a TOC
}

