\name{section.names.for.unit}
\alias{section.names.for.unit}
\title{Lists the section names in a unit}
\usage{
  section.names.for.unit(unit.id, toc)
}
\arguments{
  \item{unit.id}{the unit ID from which the sections are to
  be draw}

  \item{toc}{the TOC object to query}
}
\value{
  a list of the unique section names in the TOC within the
  specified unit
}
\description{
  This function lists the unique section names in a TOC
}

