\name{skills.for.section}
\alias{skills.for.section}
\title{Lists the skills in a section}
\usage{
  skills.for.section(unit.id, section.id, toc)
}
\arguments{
  \item{toc}{the TOC object to query}
}
\value{
  a list of the unique unit IDs in the TOC
}
\description{
  This function lists the unique unit.ids in a TOC
}

